Switch from testunit to rspec.
