+++ CHANGES.md
@@ -1,3 +1,4 @@
+## 1.1.0 - 4-May-2022
 ## 1.0.5 - 26-Oct-2019
 * Now assumes Ruby 2.
 * Updated cert.
